chore(rebuild): rebuild code structure refactoring and cleanup
